requirements.txt
argparse
csv
dnspython
requests
json
base64
urllib3
